fix:  the Bottom Navigation Bar (used on mobile) and the Mobile Top Bar lacked the necessary listeners to update their UI when switching between admin and professional mode.
